requirements.txt
alembic==0.8.10
click==6.7
dnspython==1.15.0
docker==2.1.0
docker-pycreds==0.2.1
Flask==0.12.1
Flask-HTTPAuth==3.2.2
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
MicroFlack-Common==0.1
PyJWT==1.4.2
python-dotenv==0.6.4
python-editor==1.0.3
python-etcd==0.4.5
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
urllib3==1.20
websocket-client==0.40.0
Werkzeug==0.12.1
